python/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pytesseract>=0.3.10
opencv-python-headless>=4.8.0
pdf2image>=1.16.3
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
playwright>=1.40.0
pydantic>=2.5.0
